crypto/tinycrypt: fix undefined behavior in CTR mode counter assembly
Cast nonce bytes to uint32_t before left shift to avoid
signed integer overflow when byte value >= 128. This fixes
UBSan runtime error: "left shift of N by 24 places cannot be
represented in type 'int'".

+++ lib/source/ctr_mode.c
@@ -58,8 +58,9 @@ int tc_ctr_mode(uint8_t *out, unsigned int outlen, const uint8_t *in,
 	(void)_copy(nonce, sizeof(nonce), ctr, sizeof(nonce));
 
 	/* select the last 4 bytes of the nonce to be incremented */
-	block_num = (nonce[12] << 24) | (nonce[13] << 16) |
-		    (nonce[14] << 8) | (nonce[15]);
+	block_num = ((uint32_t)nonce[12] << 24) |
+	            ((uint32_t)nonce[13] << 16) |
+	            ((uint32_t)nonce[14] << 8)  | (nonce[15]);
 	for (i = 0; i < inlen; ++i) {
 		if ((i % (TC_AES_BLOCK_SIZE)) == 0) {
 			/* encrypt data using the current nonce */
